backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# OpenAI
openai>=1.12.0

# Google Gemini (free embeddings)
google-generativeai>=0.3.0

# Vector database
qdrant-client>=1.7.0

# PostgreSQL
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
asyncpg>=0.29.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment and utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Markdown processing
markdown>=3.5.0

# HTTP client
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
